services/transaction-producer/requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.0
faker==22.6.0
